+++ core/module/theme/view/admin/admin.js.php
@@ -46,3 +46,13 @@ $("input, select").on("change", function() {
 		.appendTo("head");
 
 });
+
+/**
+ * Confirmation de réinitialisation
+ */
+$("#configAdminReset").on("click", function() {
+	var _this = $(this);
+	return core.confirm("Êtes-vous sûr de vouloir réinitialiser à son état d'origine le thème de l\'administration ?", function() {
+		$(location).attr("href", _this.attr("href"));
+	});
+});